+++ doc/release_notes.rst
@@ -20,6 +20,8 @@ Upcoming Release
 
 * Files extracted from sector-coupled data bundle have been moved from ``data/`` to ``data/sector-bundle``.
 
+* New feature multi-decade optimisation with  perfect foresight.
+
 * It is now possible to specify years for biomass potentials which do not exist
   in the JRC-ENSPRESO database, e.g. 2037. These are linearly interpolated.
 
@@ -31,6 +33,7 @@ Upcoming Release
 **Bugs and Compatibility**
 
 * A bug preventing custom powerplants specified in ``data/custom_powerplants.csv`` was fixed. (https://github.com/PyPSA/pypsa-eur/pull/732)
+
 
 PyPSA-Eur 0.8.1 (27th July 2023)
 ================================